Fix delete Slider

+++ app/Http/Controllers/admin/SliderController.php
@@ -172,7 +172,11 @@ class SliderController extends Controller
     public function destroy($id)
     {
         $slider = Slider::find($id);
-        unlink('uploads/slider/'. $slider->image);
+        
+        if(file_exists('uploads/slider/'.$slider->image)){
+            unlink('uploads/slider/'. $slider->image);
+        }
+        
         $slider->delete();
 
         return redirect()->back()->with('successMsg', 'Slider supprimer avec succes');
